Working 64-bit bitsliced sbox

+++ bs/sbox-forward.h
@@ -1,3 +1,3 @@
 #include <stdint.h>
 
-void sbox_forward(const uint32_t *inputs, uint32_t *outputs);
+void sbox_forward(const uint64_t *inputs, uint64_t *outputs);
